refactor: move early logger to sdk/packages/core/src/logging/early-logger.ts
